feat(ui): Information header for selected resource panel (#5557)
* Add new resource header for resource info panel
