+++ ui/src/app/applications/components/application-details/application-details.tsx
@@ -1,32 +1,25 @@
-import {Checkbox as ArgoCheckbox, DropDownMenu, NotificationType, SlidingPanel, Tab, Tabs, TopBarFilter} from 'argo-ui';
+import {Checkbox as ArgoCheckbox, DropDownMenu, NotificationType, SlidingPanel, TopBarFilter} from 'argo-ui';
 import * as classNames from 'classnames';
 import * as PropTypes from 'prop-types';
 import * as React from 'react';
 import {RouteComponentProps} from 'react-router';
 import {BehaviorSubject, Observable} from 'rxjs';
 
-import {DataLoader, EmptyState, ErrorNotification, EventsList, ObservableQuery, Page, Paginate, Revision, Timestamp, YamlEditor} from '../../../shared/components';
-import {AppContext} from '../../../shared/context';
+import {DataLoader, EmptyState, ErrorNotification, ObservableQuery, Page, Paginate, Revision, Timestamp} from '../../../shared/components';
+import {AppContext, ContextApis} from '../../../shared/context';
 import * as appModels from '../../../shared/models';
 import {AppDetailsPreferences, AppsDetailsViewType, services} from '../../../shared/services';
 
 import {ApplicationConditions} from '../application-conditions/application-conditions';
 import {ApplicationDeploymentHistory} from '../application-deployment-history/application-deployment-history';
-import {ApplicationNodeInfo} from '../application-node-info/application-node-info';
 import {ApplicationOperationState} from '../application-operation-state/application-operation-state';
-import {ApplicationParameters} from '../application-parameters/application-parameters';
 import {PodView} from '../application-pod-view/pod-view';
-import {ApplicationResourceEvents} from '../application-resource-events/application-resource-events';
 import {ApplicationResourceTree, ResourceTreeNode} from '../application-resource-tree/application-resource-tree';
-import {ApplicationResourcesDiff} from '../application-resources-diff/application-resources-diff';
 import {ApplicationStatusPanel} from '../application-status-panel/application-status-panel';
-import {ApplicationSummary} from '../application-summary/application-summary';
 import {ApplicationSyncPanel} from '../application-sync-panel/application-sync-panel';
-import {PodsLogsViewer} from '../pod-logs-viewer/pod-logs-viewer';
+import {ResourceDetails} from '../resource-details/resource-details';
 import * as AppUtils from '../utils';
 import {ApplicationResourceList} from './application-resource-list';
-
-const jsonMergePatch = require('json-merge-patch');
 
 require('./application-details.scss');
 
@@ -34,6 +27,21 @@ interface ApplicationDetailsState {
     page: number;
     revision?: string;
 }
+
+export const NodeInfo = (node?: string): {key: string; container: number} => {
+    const nodeContainer = {key: '', container: 0};
+    if (node) {
+        const parts = node.split('/');
+        nodeContainer.key = parts.slice(0, 4).join('/');
+        nodeContainer.container = parseInt(parts[4] || '0', 10);
+    }
+    return nodeContainer;
+};
+
+export const SelectNode = (fullName: string, containerIndex = 0, tab: string = null, appContext: ContextApis) => {
+    const node = fullName ? `${fullName}/${containerIndex}` : null;
+    appContext.navigation.goto('.', {node, tab});
+};
 
 export class ApplicationDetails extends React.Component<RouteComponentProps<{name: string}>, ApplicationDetailsState> {
     public static contextTypes = {
@@ -59,23 +67,8 @@ export class ApplicationDetails extends React.Component<RouteComponentProps<{nam
         return parseInt(new URLSearchParams(this.props.history.location.search).get('rollback'), 10);
     }
 
-    private get page(): number {
-        return parseInt(new URLSearchParams(this.props.history.location.search).get('page'), 10) || 0;
-    }
-
-    private get untilTimes(): string[] {
-        return (new URLSearchParams(this.props.history.location.search).get('untilTimes') || '').split(',') || [];
-    }
-
     private get selectedNodeInfo() {
-        const nodeContainer = {key: '', container: 0};
-        const node = new URLSearchParams(this.props.location.search).get('node');
-        if (node) {
-            const parts = node.split('/');
-            nodeContainer.key = parts.slice(0, 4).join('/');
-            nodeContainer.container = parseInt(parts[4] || '0', 10);
-        }
-        return nodeContainer;
+        return NodeInfo(new URLSearchParams(this.props.history.location.search).get('node'));
     }
 
     private get selectedNodeKey() {
@@ -157,7 +150,6 @@ export class ApplicationDetails extends React.Component<RouteComponentProps<{nam
                             const operationState = application.status.operationState;
                             const conditions = application.status.conditions || [];
                             const syncResourceKey = new URLSearchParams(this.props.history.location.search).get('deploy');
-                            const tab = new URLSearchParams(this.props.history.location.search).get('tab');
                             const filteredRes = application.status.resources.filter(res => {
                                 const resNode: ResourceTreeNode = {...res, root: null, info: null, parentRefs: [], resourceVersion: '', uid: ''};
                                 resNode.root = resNode;
@@ -299,150 +291,13 @@ export class ApplicationDetails extends React.Component<RouteComponentProps<{nam
                                                 )}
                                         </div>
                                         <SlidingPanel isShown={selectedNode != null || isAppSelected} onClose={() => this.selectNode('')}>
-                                            <div>
-                                                {selectedNode && (
-                                                    <DataLoader
-                                                        noLoaderOnInputChange={true}
-                                                        input={selectedNode.resourceVersion}
-                                                        load={async () => {
-                                                            const managedResources = await services.applications.managedResources(application.metadata.name, {
-                                                                id: {
-                                                                    name: selectedNode.name,
-                                                                    namespace: selectedNode.namespace,
-                                                                    kind: selectedNode.kind,
-                                                                    group: selectedNode.group
-                                                                }
-                                                            });
-                                                            const controlled = managedResources.find(item => AppUtils.isSameNode(selectedNode, item));
-                                                            const summary = application.status.resources.find(item => AppUtils.isSameNode(selectedNode, item));
-                                                            const controlledState = (controlled && summary && {summary, state: controlled}) || null;
-                                                            const resQuery = {...selectedNode};
-                                                            if (controlled && controlled.targetState) {
-                                                                resQuery.version = AppUtils.parseApiVersion(controlled.targetState.apiVersion).version;
-                                                            }
-                                                            const liveState = await services.applications.getResource(application.metadata.name, resQuery).catch(() => null);
-                                                            const events =
-                                                                (liveState &&
-                                                                    (await services.applications.resourceEvents(application.metadata.name, {
-                                                                        name: liveState.metadata.name,
-                                                                        namespace: liveState.metadata.namespace,
-                                                                        uid: liveState.metadata.uid
-                                                                    }))) ||
-                                                                [];
-                                                            let podState: appModels.State;
-                                                            if (selectedNode.kind === 'Pod') {
-                                                                podState = liveState;
-                                                            } else {
-                                                                const childPod = AppUtils.findChildPod(selectedNode, tree);
-                                                                if (childPod) {
-                                                                    podState = await services.applications.getResource(application.metadata.name, childPod).catch(() => null);
-                                                                }
-                                                            }
-
-                                                            return {controlledState, liveState, events, podState};
-                                                        }}>
-                                                        {data => (
-                                                            <Tabs
-                                                                navTransparent={true}
-                                                                tabs={this.getResourceTabs(application, selectedNode, data.liveState, data.podState, data.events, [
-                                                                    {
-                                                                        title: 'SUMMARY',
-                                                                        key: 'summary',
-                                                                        content: (
-                                                                            <ApplicationNodeInfo
-                                                                                application={application}
-                                                                                live={data.liveState}
-                                                                                controlled={data.controlledState}
-                                                                                node={selectedNode}
-                                                                            />
-                                                                        )
-                                                                    }
-                                                                ])}
-                                                                selectedTabKey={tab}
-                                                                onTabSelected={selected => this.appContext.apis.navigation.goto('.', {tab: selected})}
-                                                            />
-                                                        )}
-                                                    </DataLoader>
-                                                )}
-                                                {isAppSelected && (
-                                                    <Tabs
-                                                        navTransparent={true}
-                                                        tabs={[
-                                                            {
-                                                                title: 'SUMMARY',
-                                                                key: 'summary',
-                                                                content: <ApplicationSummary app={application} updateApp={app => this.updateApp(app)} />
-                                                            },
-                                                            {
-                                                                title: 'PARAMETERS',
-                                                                key: 'parameters',
-                                                                content: (
-                                                                    <DataLoader
-                                                                        key='appDetails'
-                                                                        input={application}
-                                                                        load={app =>
-                                                                            services.repos.appDetails(app.spec.source, app.metadata.name).catch(() => ({
-                                                                                type: 'Directory' as appModels.AppSourceType,
-                                                                                path: application.spec.source.path
-                                                                            }))
-                                                                        }>
-                                                                        {(details: appModels.RepoAppDetails) => (
-                                                                            <ApplicationParameters save={app => this.updateApp(app)} application={application} details={details} />
-                                                                        )}
-                                                                    </DataLoader>
-                                                                )
-                                                            },
-                                                            {
-                                                                title: 'MANIFEST',
-                                                                key: 'manifest',
-                                                                content: (
-                                                                    <YamlEditor
-                                                                        minHeight={800}
-                                                                        input={application.spec}
-                                                                        onSave={async patch => {
-                                                                            const spec = JSON.parse(JSON.stringify(application.spec));
-                                                                            return services.applications.updateSpec(
-                                                                                application.metadata.name,
-                                                                                jsonMergePatch.apply(spec, JSON.parse(patch))
-                                                                            );
-                                                                        }}
-                                                                    />
-                                                                )
-                                                            },
-                                                            {
-                                                                icon: 'fa fa-file-medical',
-                                                                title: 'DIFF',
-                                                                key: 'diff',
-                                                                content: (
-                                                                    <DataLoader
-                                                                        key='diff'
-                                                                        load={async () =>
-                                                                            await services.applications.managedResources(application.metadata.name, {
-                                                                                fields: [
-                                                                                    'items.normalizedLiveState',
-                                                                                    'items.predictedLiveState',
-                                                                                    'items.group',
-                                                                                    'items.kind',
-                                                                                    'items.namespace',
-                                                                                    'items.name'
-                                                                                ]
-                                                                            })
-                                                                        }>
-                                                                        {managedResources => <ApplicationResourcesDiff states={managedResources} />}
-                                                                    </DataLoader>
-                                                                )
-                                                            },
-                                                            {
-                                                                title: 'EVENTS',
-                                                                key: 'event',
-                                                                content: <ApplicationResourceEvents applicationName={application.metadata.name} />
-                                                            }
-                                                        ]}
-                                                        selectedTabKey={tab}
-                                                        onTabSelected={selected => this.appContext.apis.navigation.goto('.', {tab: selected})}
-                                                    />
-                                                )}
-                                            </div>
+                                            <ResourceDetails
+                                                tree={tree}
+                                                application={application}
+                                                isAppSelected={isAppSelected}
+                                                updateApp={app => this.updateApp(app)}
+                                                selectedNode={selectedNode}
+                                            />
                                         </SlidingPanel>
                                         <ApplicationSyncPanel
                                             application={application}
@@ -694,8 +549,7 @@ export class ApplicationDetails extends React.Component<RouteComponentProps<{nam
     }
 
     private selectNode(fullName: string, containerIndex = 0, tab: string = null) {
-        const node = fullName ? `${fullName}/${containerIndex}` : null;
-        this.appContext.apis.navigation.goto('.', {node, tab});
+        SelectNode(fullName, containerIndex, tab, this.appContext.apis);
     }
 
     private async rollbackApplication(revisionHistory: appModels.RevisionHistory, application: appModels.Application) {
@@ -732,84 +586,5 @@ Are you sure you want to disable auto-sync and rollback application '${this.prop
 
     private async deleteApplication() {
         await AppUtils.deleteApplication(this.props.match.params.name, this.appContext.apis);
-    }
-
-    private getResourceTabs(
-        application: appModels.Application,
-        node: appModels.ResourceNode,
-        state: appModels.State,
-        podState: appModels.State,
-        events: appModels.Event[],
-        tabs: Tab[]
-    ) {
-        if (state) {
-            const numErrors = events.filter(event => event.type !== 'Normal').reduce((total, event) => total + event.count, 0);
-            tabs.push({
-                title: 'EVENTS',
-                badge: (numErrors > 0 && numErrors) || null,
-                key: 'events',
-                content: (
-                    <div className='application-resource-events'>
-                        <EventsList events={events} />
-                    </div>
-                )
-            });
-        }
-        if (podState) {
-            const containerGroups = [
-                {
-                    offset: 0,
-                    title: 'CONTAINERS',
-                    containers: podState.spec.containers || []
-                },
-                {
-                    offset: (podState.spec.containers || []).length,
-                    title: 'INIT CONTAINERS',
-                    containers: podState.spec.initContainers || []
-                }
-            ];
-            tabs = tabs.concat([
-                {
-                    key: 'logs',
-                    title: 'LOGS',
-                    content: (
-                        <div className='application-details__tab-content-full-height'>
-                            <div className='row'>
-                                <div className='columns small-3 medium-2'>
-                                    {containerGroups.map(group => (
-                                        <div key={group.title} style={{marginBottom: '1em'}}>
-                                            {group.containers.length > 0 && <p>{group.title}</p>}
-                                            {group.containers.map((container: any, i: number) => (
-                                                <div
-                                                    className='application-details__container'
-                                                    key={container.name}
-                                                    onClick={() => this.selectNode(this.selectedNodeKey, group.offset + i, 'logs')}>
-                                                    {group.offset + i === this.selectedNodeInfo.container && <i className='fa fa-angle-right' />}
-                                                    <span title={container.name}>{container.name}</span>
-                                                </div>
-                                            ))}
-                                        </div>
-                                    ))}
-                                </div>
-                                <div className='columns small-9 medium-10'>
-                                    <PodsLogsViewer
-                                        podName={(state.kind === 'Pod' && state.metadata.name) || ''}
-                                        group={node.group}
-                                        kind={node.kind}
-                                        name={node.name}
-                                        namespace={podState.metadata.namespace}
-                                        applicationName={application.metadata.name}
-                                        containerName={AppUtils.getContainerName(podState, this.selectedNodeInfo.container)}
-                                        page={{number: this.page, untilTimes: this.untilTimes}}
-                                        setPage={pageData => this.appContext.apis.navigation.goto('.', {page: pageData.number, untilTimes: pageData.untilTimes.join(',')})}
-                                    />
-                                </div>
-                            </div>
-                        </div>
-                    )
-                }
-            ]);
-        }
-        return tabs;
     }
 }
