+++ ui/src/app/applications/components/resource-details/resource-details.scss
@@ -1,0 +1,17 @@
+@import 'node_modules/argo-ui/src/styles/config';
+
+.resource-details {
+    &__header {
+        display: flex;
+        margin-bottom: 1em;
+        align-items: center;
+        padding-right: 48px;
+        h1 {
+            color: $argo-color-gray-8;
+            font-size: 23px;
+            font-weight: 500;
+            margin-right: 5px;
+            margin-bottom: 0;
+        }
+    }
+}
